+++ flutter_inappwebview_platform_interface/CHANGELOG.md
@@ -2,8 +2,9 @@
 
 - Added `isClassSupported`, `isPropertySupported`, `isMethodSupported` static methods for all main classes, such as `PlatformInAppWebViewController`, `InAppWebViewSettings`, `PlatformInAppBrowser`, etc., in order to check if a class, property, or method is supported by the platform at runtime
 - Added `isSupported` method to all custom enum classes
-- Added `saveState`, `restoreState` methods to `PlatformInAppWebViewController` class
-- Added `useOnAjaxReadyStateChange`, `useOnAjaxProgress`, `useOnShowFileChooser`, `corsAllowlist` properties to `InAppWebViewSettings`
+- Added `saveState`, `restoreState`, `requestEnterFullscreen`, `requestExitFullscreen`, `setVisible`, `setTargetRefreshRate`, `getTargetRefreshRate`, `requestPointerLock`, `requestPointerUnlock` methods to `PlatformInAppWebViewController` class
+- Added `useOnAjaxReadyStateChange`, `useOnAjaxProgress`, `useOnShowFileChooser`, `corsAllowlist`, `itpEnabled` properties to `InAppWebViewSettings`
+- Added `onShowFileChooser` WebView events
 - Update code documentation
 
 ## 1.4.0-beta.2